Move utils & pass lint
